added parameters support. added some unit tests
