+++ src/ast/enrichment/annotate_project_all.h
@@ -1,0 +1,11 @@
+/*
+* Copyright 2018-2019 Redis Labs Ltd. and Contributors
+*
+* This file is available under the Redis Labs Source Available License Agreement
+*/
+
+#pragma once
+
+#include "../ast.h"
+
+void AST_AnnotateProjectAll(AST *ast);
