+++ src/ast/enrichment/annotate_projected_named_paths.h
@@ -1,0 +1,11 @@
+/*
+* Copyright 2018-2019 Redis Labs Ltd. and Contributors
+*
+* This file is available under the Redis Labs Source Available License Agreement
+*/
+
+#pragma once
+
+#include "../ast.h"
+
+void AST_AnnotateNamedPaths(AST *ast);
